feat: ensure cart product can be decreased

+++ bewear/src/actions/decrease-cart-quantity/index.ts
@@ -10,10 +10,10 @@ import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
 
-export const removeProductFromCart = async (data: z.infer<typeof removeProductFromCartSchema>) => {
-    removeProductFromCartSchema.parse(data);
+export const decreaseCartProductQuantity = async (data: z.infer<typeof decreaseCartProductQuantitySchema>) => {
+    decreaseCartProductQuantitySchema.parse(data);
     const session = await auth.api.getSession({
         headers: await headers(),
     });
@@ -33,5 +33,14 @@ export const removeProductFromCart = async (data: z.infer<typeof removeProductFr
     if (cartDoesNotBelongToUser) {
         throw new Error("You don't have permission to remove this product from the cart");
     }
-    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+
+    if (cartItem.quantity === 1) {
+        await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+        return;
+    }
+
+    await db.
+        update(cartItemTable)
+        .set({ quantity: cartItem.quantity - 1 })
+        .where(eq(cartItemTable.id, cartItem.id));
 };
